change (types, springbone): rename node collider to spring bone collider (1)
