+++ packages/types-vrmc-springbone-1.0/src/SpringBoneCollider.ts
@@ -1,4 +1,4 @@
-import type { NodeColliderShape } from './NodeColliderShape';
+import type { NodeColliderShape } from './SpringBoneColliderShape';
 
 /**
  * collider definition for SpringBone
