+++ packages/types-vrmc-springbone-1.0/src/SpringBoneColliderCapsule.ts
