+++ packages/types-vrmc-springbone-1.0/src/index.ts
@@ -1,3 +1,7 @@
 export * from './SpringBone';
+export * from './SpringBoneCollider';
+export * from './SpringBoneColliderCapsule';
+export * from './SpringBoneColliderShape';
+export * from './SpringBoneColliderSphere';
 export * from './SpringBoneJoint';
 export * from './SpringBoneSpring';
